Move graphql server info to index
