+++ Server/src/config/middlewares.js
@@ -1,16 +1,6 @@
 import bodyParser from "body-parser";
-import { graphiqlExpress, graphqlExpress } from "apollo-server-express";
-import { makeExecutableSchema } from "graphql-tools";
 
-import constants from "./constants";
-import typeDefs from "../graphql/schema";
-import resolvers from "../graphql/resolvers";
 import { decodeToken } from "../services/auth";
-
-const schema = makeExecutableSchema({
-  typeDefs,
-  resolvers
-});
 
 async function auth(req, res, next) {
   try {
@@ -30,19 +20,5 @@ async function auth(req, res, next) {
 export default app => {
   app.use(bodyParser.json());
   app.use(auth);
-  app.use(
-    "/graphiql",
-    graphiqlExpress({
-      endpointURL: constants.GRAPHQL_PATH
-    })
-  );
-  app.use(
-    constants.GRAPHQL_PATH,
-    graphqlExpress(req => ({
-      schema,
-      context: {
-        user: req.user
-      }
-    }))
-  );
+
 };
