+++ Server/src/index.js
@@ -1,9 +1,16 @@
 /* eslint-disable no-console */
 
 import express from "express";
+
 import { createServer } from "http";
+import { graphiqlExpress, graphqlExpress } from "apollo-server-express";
+import { makeExecutableSchema } from "graphql-tools";
+import { SubscriptionServer } from 'subscriptions-transport-ws'
+import { execute, subscribe } from 'graphql'
 
 import "./config/db";
+import typeDefs from "./graphql/schema";
+import resolvers from "./graphql/resolvers";
 import constants from "./config/constants";
 import middlewares from "./config/middlewares";
 import mocks from "./mocks";
@@ -12,6 +19,29 @@ const app = express();
 
 middlewares(app);
 
+app.use(
+  "/graphiql",
+  graphiqlExpress({
+    endpointURL: constants.GRAPHQL_PATH,
+    SubscriptionsEndpoint: `ws:/localhost:${constants.PORT}${constants.SUBSCRIPTIONS_PATH}`
+  })
+);
+
+const schema = makeExecutableSchema({
+  typeDefs,
+  resolvers
+});
+
+app.use(
+  constants.GRAPHQL_PATH,
+  graphqlExpress(req => ({
+    schema,
+    context: {
+      user: req.user
+    }
+  })),
+);
+
 const graphQLServer = createServer(app);
 
 mocks().then(() => {
@@ -19,6 +49,14 @@ mocks().then(() => {
     if (err) {
       console.error(err);
     } else {
+      new SubscriptionServer({
+        schema,
+        execute,
+        subscribe
+      }, {
+        server: graphQLServer,
+        path: constants.SUBSCRIPTIONS_PATH
+      })
       console.log(`App listen to port: ${constants.PORT}`);
     }
   });
